requirements/base.txt
Django==2.2.10
django-filter==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.0
Markdown==3.2
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
Scrapy==1.8.0
